+++ e2e/src/object/room.rs
@@ -336,6 +336,7 @@ impl Object<Room> {
     /// # Errors
     ///
     /// If failed to execute JS statement.
+    #[expect(clippy::too_many_lines, reason = "still easy to follow")]
     pub async fn connections_store(
         &self,
     ) -> Result<Object<ConnectionStore>, Error> {
@@ -356,10 +357,14 @@ impl Object<Room> {
                         tracks: [],
                         subs: []
                     };
+                    let stateListener = {
+                        subs: [],
+                    };
                     let connection = {
                         conn: conn,
                         tracksStore: tracksStore,
                         closeListener: closeListener,
+                        stateListener: stateListener,
                     };
                     conn.on_remote_track_added((t) => {
                         let track = {
@@ -418,6 +423,11 @@ impl Object<Room> {
                                 return sub(track);
                             });
                         tracksStore.subs = newStoreSubs;
+                    });
+                    conn.on_state_change((state) => {
+                        for (const sub of stateListener.subs) {
+                            sub(state);
+                        }
                     });
                     conn.on_close(() => {
                         closeListener.isClosed = true;